+++ src/wl_collocation.py
@@ -1204,7 +1204,7 @@ def generate_table(main, table):
             thread_collocation_table = wl_threading.Wl_Thread(worker_collocation_table)
             thread_collocation_table.start_worker()
         else:
-            wl_msg_box.wl_msg_box_missing_search_term_optional(main)
+            wl_msg_box.wl_msg_box_missing_search_terms_optional(main)
 
             wl_msg.wl_msg_generate_table_error(main)
     else:
@@ -1352,7 +1352,7 @@ def generate_fig(main):
             thread_collocation_fig = wl_threading.Wl_Thread(worker_collocation_fig)
             thread_collocation_fig.start_worker()
         else:
-            wl_msg_box.wl_msg_box_missing_search_term_optional(main)
+            wl_msg_box.wl_msg_box_missing_search_terms_optional(main)
 
             wl_msg.wl_msg_generate_fig_error(main)
     else: